moved api calls to the context and cleaned up app.js
