+++ src/App.js
@@ -1,8 +1,5 @@
 import "./style.scss";
 import { Routes, Route } from "react-router-dom";
-import { useEffect, useState } from "react";
-import axios from "axios";
-import { apiClient } from "./Api/axios.instance";
 import Home from "./pages/Home/Home";
 import ProductListing from "./pages/ProductListing/ProductListing";
 import ProductDetails from "./pages/ProductDetails/ProductDetails";
@@ -19,39 +16,7 @@ import PrivateRoute from "./PrivateRoutes/PrivateRoute";
 
 function App() {
   const { user } = useAuth();
-  const { dispatch } = useProduct();
-  const [loading, setLoading] = useState("");
-
-  useEffect(() => {
-    const fetchProductsAndCategories = async () => {
-      setLoading(true);
-      const products = apiClient.get("/products");
-      const categories = apiClient.get("/category");
-
-      const response = await axios.all([products, categories]);
-
-      dispatch({ type: "GET_PRODUCTS", payload: response[0].data.products });
-      dispatch({ type: "GET_CATEGORIES", payload: response[1].data.category });
-      setLoading(false);
-    };
-    fetchProductsAndCategories();
-  }, [user, dispatch]);
-
-  useEffect(() => {
-    const fetchCartAndWishlist = async () => {
-      const fetchCart = apiClient.get("/cart");
-      const fetchWishlist = apiClient.get("/wishlist");
-
-      const response = await axios.all([fetchCart, fetchWishlist]);
-
-      dispatch({ type: "SET_CART", payload: response[0].data.cart.cartItems });
-      dispatch({
-        type: "SET_WISHLIST",
-        payload: response[1].data.wishlist.wishlistItems
-      });
-    };
-    user && fetchCartAndWishlist();
-  }, [dispatch, user]);
+  const { loading } = useProduct();
 
   return (
     <div className="App-container">
